requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
idna==3.4
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
stripe==7.1.0
typing_extensions==4.8.0
urllib3==2.0.7
